requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
defusedxml==0.7.1
Django==4.0
django-allauth==0.46.0
django-crispy-forms==1.13.0
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
whitenoise==5.3.0

